pyproject.toml
[project]
name = "ticket-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "fastembed>=0.7.3",
    "numpy>=2.3.3",
    "openai>=2.1.0",
    "python-decouple>=3.8",
    "qdrant-client>=1.15.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "uvicorn>=0.37.0",
]
